+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.98.33", date: "2026-04-04" };
+  return { version: "8.98.34", date: "2026-04-04" };
 }